refactor : RandomRangeGenerator, RandomRangeGeneratorImpl 패키지 이동
